internal/provider: enhance schema apply with approval flow (#188)
This pull request adds an approval flow when applying `atlas_schema`
resources. Based on the lint policy, the process will exit and wait for
approval before continuing, depending on the severity level you set

### Usage

- To enable the review flow, set the `lint.review` option to one of the
following values: `ALWAYS`, `ERROR`, or `WARNING`:

  - `ALWAYS`: Triggers an approval plan for every schema change.
- `ERROR`/`WARNING`: Triggers an approval plan only when the schema
change violates the corresponding lint policy.

```hcl
resource "atlas_schema" "db" {
  lint {
    review = "ALWAYS" //  "ALWAYS", "ERROR", "WARNING"
  }
}
```

- To control the timeout of approval process, set `lint.review_timeout`.
Default to `0s`

```hcl
resource "atlas_schema" "db" {
  lint {
    review = "ALWAYS"
    review_timeout = "1m"
  }
}
```

### Demo

When `lint.review` is enabled, the provider will automatically create a
plan and require user approval before proceeding. If `review_timeout` is
not set, the process will throw an error immediately:

![Screenshot 2025-04-14 at 13 22
09](https://github.com/user-attachments/assets/622108f1-1e18-4c7c-bd2c-bdf06c52f7c5)


https://github.com/user-attachments/assets/dc92e6d6-6433-4f59-a6c9-5d9c8c856ba1

If `lint.review_timeout` is set, the process will wait for the plan to
be approved before continuing to apply the schema.


https://github.com/user-attachments/assets/e75dd1df-1ed8-4892-92b1-8ad698572f91

If `lint.review_timeout` is set and the timeout is reached before
approval, an error will be displayed:

![Screenshot 2025-04-14 at 14 24
31](https://github.com/user-attachments/assets/b28507d8-7706-42d3-b1d2-d9fe4b97981b)